Add files via upload
This repository contains a Python script that scrapes the member list of the Bangladesh Garment Manufacturers and Exporters Association (BGMEA) from their official website. It uses Selenium, Requests, BeautifulSoup, and Pandas to extract and store the data in a CSV file.

Motivation
The motivation for this project was to collect data on the BGMEA members for research purposes. The BGMEA website provides a list of its members with some basic information such as name, registration number, contact person, and email address. However, it does not provide other information such as address, phone number, fax number, product type, export country, etc. These information are available on individual member pages that can be accessed by clicking on each member name. However, manually visiting each page and copying the data is tedious and time-consuming. Therefore, I decided to write a script that automates this process and scrapes all the relevant data from both the list page and the individual pages.

How it works
The script works as follows:

It defines an empty DataFrame to store the data with columns for Member/Company Name, BGMEA Reg No, Contact Person, Email Address, Mobile, Address, Phone Number, Fax Number, Product Type, Export Country.
It defines a base URL for the list page with a page parameter that can be changed to access different pages.
It defines a number of pages to scrape. The default value is 370, which corresponds to the total number of pages on the BGMEA website as of September 2023. You can change this value to any number you want.
It loops over the page numbers and appends the page number to the base URL to get the full URL for each page.
It uses Selenium to get the HTML content of each page and waits for the table element to load using explicit wait and CSS selector.
It loops through each row in the table and extracts the text from each cell using XPath locator. It also constructs a URL for each individual member page by appending the row index to a base URL for individual pages.
It uses Requests and BeautifulSoup to get the HTML content of each individual page and parses it using a table element with a class name of table-bordered.
It loops through each row in the table element and extracts the text from each cell. It also checks if there are four cells in each row as some rows have only two cells with merged columns.
It appends all the extracted data to the DataFrame as a new row.
It handles exceptions such as missing table element or timeout error using try-except blocks and prints an error message if any exception occurs.
It saves the DataFrame as a CSV file with an output file name of out.csv.
Dependencies
To run this script, you need to have Python 3.x installed on your machine. You also need to install the following libraries using pip:

Selenium
Requests
BeautifulSoup
Pandas
You also need to have a Chrome driver installed on your machine and added to your PATH variable. You can download it from here.

Usage
To use this script, follow these steps:

Clone or download this repository to your local machine.
Install the required libraries using pip install -r requirements.txt.
Open the scraper.py file and change the num_pages variable to the number of pages you want to scrape. The default value is 370.
Run the script using python scraper.py.
Wait for the script to finish scraping and check the out.csv file for the output.
Limitations
This script has some limitations that you should be aware of:

It may not work if the BGMEA website changes its layout, structure, or URL format.
It may take a long time to scrape all the pages as it has to open each individual page and wait for it to load.
It may encounter some errors or inconsistencies in the data such as missing values, typos, or duplicates.
It may violate the terms and conditions of the BGMEA website and cause some legal issues if used for commercial or malicious purposes.
License
This script is licensed under the MIT License. See the LICENSE file for more details.